coverage_web_system/requirements.txt
Django==2.1
channels==2.1.2
daphne==2.2.1
asgiref==2.3.2
celery==4.2.1
django_eventstream==2.1.0
inotify==0.2.10
gitpython==2.1.10
redis==2.10.6
asgi_redis==1.4.3
django-filebrowser==3.10.1
django-filebrowser-no-grappelli==3.7.5
image==1.5.27
pillow==5.3.0